Change from 32-bit to 64-bit integer vertices

+++ app/src/main/java/navigation/Graph.java
@@ -10,7 +10,7 @@ import java.util.Map;
 class Graph implements Serializable {
     private final int vertexCount;
     // Use a hashmap to support non-index values.
-    private final Map<Integer, List<Edge>> adjacentEdges;
+    private final Map<Long, List<Edge>> adjacentEdges;
     private int edgeCount;
 
     public Graph(int vertexCount) {
